feat: add solution modules/cjsToEsm.cjs
